article rework and fixes, late values moved to init everywhere, renderbox error resolved on userpage
