+++ lib/homePage.dart
@@ -27,7 +27,7 @@ class HomePage extends StatefulWidget {
 
 class _HomePageState extends State<HomePage>
     with SingleTickerProviderStateMixin {
-  late Future<List<Item>>? items;
+  late Future<List<Item>?> items;
 
   late AnimationController animationController;
 
@@ -36,63 +36,61 @@ class _HomePageState extends State<HomePage>
     animationController =
         AnimationController(vsync: this, duration: Duration(milliseconds: 250));
     animationController.addListener(() => setState(() {}));
+    items = RemoteItems.getAllItems(1, itemParams: widget._filters);
 
     super.initState();
   }
 
   @override
-  Widget build(BuildContext context) {
-    items = RemoteItems.getAllItems(1, itemParams: widget._filters);
-    return Scaffold(
-        floatingActionButton: MainButton(
-            iconData: Icons.menu,
-            onPressed: () {
-              if (animationController.isCompleted) {
-                animationController.reverse();
-                volume = 0;
-              } else {
-                animationController.forward();
-                volume = 0.5;
-              }
-            }),
-        floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
-        body: Stack(children: [
-          Container(
-              color: kBlack,
-              child: FutureBuilder<List<dynamic>>(
-                  future: items,
-                  builder: (context, snapshot) =>
-                      (snapshot.connectionState == ConnectionState.waiting
-                          ? Center(child: Image.asset("assets/load.gif"))
-                          : (snapshot.hasData
-                              ? ListView.builder(
-                                  itemCount: snapshot.data!.length,
-                                  itemBuilder: (context, i) => snapshot.data![i]
-                                      .buildCard(context, widget._loggedUser))
-                              : (snapshot.hasError
-                                  ? ErrorWidgets.futureBuilderError()
-                                  : ErrorWidgets.futureBuilderEmpty()))))),
-          MainButtonClicked(buttons: [
-            SecondaryButtonData(
-                Icons.add,
-                () => Navigator.of(context).push(MaterialPageRoute(
-                    builder: (context) =>
-                        AddItem(loggedUser: widget._loggedUser)))),
-            SecondaryButtonData(
-                Icons.filter_alt,
-                () => Navigator.of(context).push(MaterialPageRoute(
-                    builder: (context) =>
-                        FilterPage(loggedUser: widget._loggedUser)))),
-            SecondaryButtonData(
-                Icons.person,
-                () => Navigator.of(context).push(MaterialPageRoute(
-                    builder: (context) =>
-                        UserPage(loggedUser: widget._loggedUser)))),
-            SecondaryButtonData(
-                Icons.article,
-                () => Navigator.of(context).push(
-                    MaterialPageRoute(builder: (context) => ArticlesScreen())))
-          ], volume: volume)
-        ]));
-  }
+  Widget build(BuildContext context) => Scaffold(
+      floatingActionButton: MainButton(
+          iconData: Icons.menu,
+          onPressed: () {
+            if (animationController.isCompleted) {
+              animationController.reverse();
+              volume = 0;
+            } else {
+              animationController.forward();
+              volume = 0.5;
+            }
+          }),
+      floatingActionButtonLocation: FloatingActionButtonLocation.endFloat,
+      body: Stack(children: [
+        Container(
+            color: kBlack,
+            child: FutureBuilder<List<Item>?>(
+                future: items,
+                builder: (context, snapshot) =>
+                    (snapshot.connectionState == ConnectionState.waiting
+                        ? Center(child: Image.asset("assets/load.gif"))
+                        : (snapshot.hasData
+                            ? ListView.builder(
+                                itemCount: snapshot.data!.length,
+                                itemBuilder: (context, i) => snapshot.data![i]
+                                    .buildCard(context, widget._loggedUser))
+                            : (snapshot.hasError
+                                ? ErrorWidgets.futureBuilderError()
+                                : ErrorWidgets.futureBuilderEmpty()))))),
+        MainButtonClicked(buttons: [
+          SecondaryButtonData(
+              Icons.add,
+              () => Navigator.of(context).push(MaterialPageRoute(
+                  builder: (context) =>
+                      AddItem(loggedUser: widget._loggedUser)))),
+          SecondaryButtonData(
+              Icons.filter_alt,
+              () => Navigator.of(context).push(MaterialPageRoute(
+                  builder: (context) =>
+                      FilterPage(loggedUser: widget._loggedUser)))),
+          SecondaryButtonData(
+              Icons.person,
+              () => Navigator.of(context).push(MaterialPageRoute(
+                  builder: (context) =>
+                      UserPage(loggedUser: widget._loggedUser)))),
+          SecondaryButtonData(
+              Icons.article,
+              () => Navigator.of(context).push(
+                  MaterialPageRoute(builder: (context) => ArticlesScreen())))
+        ], volume: volume)
+      ]));
 }
