+++ lib/remote/ratings.dart
@@ -20,7 +20,7 @@ class RemoteRatings {
     return response.statusCode == 200 && response.body != "ERROR" ? true : null;
   }
 
-  static Future<List<Rating>>? getRatings(String userId) async {
+  static Future<List<Rating>?> getRatings(String userId) async {
     final Response response = await post(
         Uri.parse(Uri.encodeFull(url + "ratingGet.php")),
         headers: {
@@ -33,6 +33,6 @@ class RemoteRatings {
             .cast()
             .map<Rating>((rating) => Rating.fromJson(rating))
             .toList()
-        : [];
+        : null;
   }
 }
